Rename protocol.py -> types.py
