+++ docs/api.md
@@ -9,7 +9,7 @@
 
 ## Custom Types
 
-::: mcbootflash.protocol
+::: mcbootflash.types
     options:
       show_root_toc_entry: false
       heading_level: 3
