+++ src/mcbootflash/util.py
@@ -5,7 +5,7 @@ from typing import Iterator
 
 import bincopy  # type: ignore[import-untyped]
 
-from mcbootflash.protocol import BootAttrs, Chunk, Command
+from mcbootflash.types import BootAttrs, Chunk, Command
 
 
 def chunked(
